services/amplifier-skills/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amplifier-skills"
version = "0.1.0"
description = "Amplifier Skills IPC service — domain knowledge loading tool"
requires-python = ">=3.11"
dependencies = [
    "amplifier-ipc-protocol",
    "pyyaml>=6.0",
]

[project.scripts]
amplifier-skills-serve = "amplifier_skills.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/amplifier_skills"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.pyright]
pythonVersion = "3.11"
extraPaths = ["src"]

[tool.uv.sources]
amplifier-ipc-protocol = { path = "../../amplifier-ipc-protocol" }
